Extract joins in a separate group
