+++ frontend/test/metabase/scenarios/joins/reproductions/13468-ui-overlay-after-joined-question-save.cy.spec.js
