+++ frontend/test/metabase/scenarios/joins/reproductions/17710-notebook-incomplete-joins-removed.cy.spec.js
