+++ frontend/test/metabase/scenarios/joins/reproductions/17712-notebook-extra-sections-removed.cy.spec.js
