+++ frontend/test/metabase/scenarios/joins/reproductions/17968-notebook-join-table-names.cy.spec.js
